more test restructured
