+++ tests/test_integration/test_backends/test_kubernetes.py
@@ -1,0 +1,20 @@
+"""
+Integration tests for Kubernetes backend functionality.
+
+Tests the KubernetesDeploymentService with real-world scenarios
+against mock Kubernetes clusters when available.
+"""
+
+import pytest
+
+pytestmark = pytest.mark.integration
+
+
+class TestKubernetesIntegration:
+    """Integration tests for Kubernetes backend."""
+
+    @pytest.mark.skip(reason="Requires Kubernetes cluster")
+    def test_full_deployment_lifecycle(self):
+        """Test complete deployment lifecycle against real cluster."""
+        # This would test against a real Kubernetes cluster (kind, minikube, etc.)
+        pass
