+++ tests/test_integration/test_core/test_new_commands.py
@@ -1,0 +1,80 @@
+"""
+End-to-end integration tests for new CLI commands.
+"""
+
+from unittest.mock import Mock, patch
+
+import pytest
+
+pytestmark = pytest.mark.integration
+
+
+class TestEndToEndScenarios:
+    """End-to-end scenario tests."""
+
+    def test_full_cleanup_scenario(self):
+        """Test a complete cleanup scenario."""
+        # This would be a real end-to-end test that actually creates and cleans containers
+        # For now, we'll mock it but structure it like a real test
+
+        with (
+            patch("subprocess.run") as mock_run,
+            patch(
+                "mcp_template.backends.docker.DockerDeploymentService._ensure_docker_available"
+            ),
+        ):
+
+            # Mock docker commands
+            mock_run.side_effect = [
+                # List containers
+                Mock(stdout="container1\tdemo_1\tExited (0)", returncode=0),
+                # Remove container
+                Mock(returncode=0),
+                # List images
+                Mock(stdout="", returncode=0),  # No dangling images
+            ]
+
+            from mcp_template.backends.docker import DockerDeploymentService
+
+            service = DockerDeploymentService()
+
+            # Test container cleanup
+            result = service.cleanup_stopped_containers()
+            assert result["success"] is True
+            assert len(result["cleaned_containers"]) == 1
+
+            # Test image cleanup
+            result = service.cleanup_dangling_images()
+            assert result["success"] is True
+            assert len(result["cleaned_images"]) == 0
+
+    @pytest.mark.skip(reason="Complex mocking scenario - covered by unit tests")
+    def test_shell_connection_scenario(self):
+        """Test a complete shell connection scenario."""
+        # This test is complex to mock properly due to os.execvp behavior
+        # The functionality is covered by unit tests in test_docker_connect.py
+        pass
+
+    def test_config_display_scenario(self):
+        """Test a complete config display scenario."""
+        from mcp_template.core.template_manager import TemplateManager
+
+        manager = TemplateManager("docker")
+
+        # Mock template discovery and schema retrieval
+        with patch.object(manager, "get_template_config_schema") as mock_schema:
+            mock_schema.return_value = {
+                "properties": {
+                    "test_prop": {
+                        "type": "string",
+                        "description": "Test property",
+                        "default": "test_value",
+                    }
+                }
+            }
+
+            schema = manager.get_template_config_schema("demo")
+
+            assert schema is not None
+            assert "properties" in schema
+            assert "test_prop" in schema["properties"]
